整合Wrapper功能到Banner中，重構
