+++ app/_components/banner.module.css
@@ -5,12 +5,20 @@
   margin: 0;
   padding: 0;
   overflow: hidden;
+  background-color: #000;
 }
 
 .banner img {
   width: 100%;
   height: auto;
   display: block;
+  object-fit: cover;
+  opacity: 0.9;
+  transition: opacity 0.3s ease;
+}
+
+.banner:hover img {
+  opacity: 0.8;
 }
 
 .banner_title {
@@ -19,8 +27,27 @@
   left: 50%;
   transform: translate(-50%, -50%);
   color: white;
-  font-size: 2.5rem;
+  font-size: clamp(1.5rem, 5vw, 2.5rem);
   font-weight: bold;
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
+  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5), -1px -1px 0 rgba(0, 0, 0, 0.7);
   z-index: 1;
+  text-align: center;
+  padding: 0.5rem 1rem;
+  user-select: none;
+  cursor: default;
+  letter-spacing: 0.1em;
+  white-space: nowrap;
+}
+
+@media (max-width: 768px) {
+  .banner_title {
+    padding: 0.3rem 0.8rem;
+    text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5);
+  }
+}
+
+@media (max-width: 480px) {
+  .banner {
+    min-height: 150px;
+  }
 }
